% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCreateCashflowMatrix.R
\name{fGetCashflowsBond}
\alias{fGetCashflowsBond}
\title{Gets the cashflow schedule for a bond}
\usage{
fGetCashflowsBond(dfInstrument)
}
\arguments{
\item{dfInstrument}{A market instrument as a dataframe with columns Frequency, Tenor and Rate with Type in (LIBOR, SWAP), Tenor the instrument maturity in years and rate the rate per annum}
}
\description{
Gets the cashflow schedule for a bond
}
