% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCreateTimeVector.R
\name{fGetTimesBond}
\alias{fGetTimesBond}
\title{Extract the payment dates of a Bond in years}
\usage{
fGetTimesBond(dfInstrument)
}
\arguments{
\item{dfInstrument}{A dataframe of an instrument with at least columns Frequency and Tenor}
}
\description{
Extract the payment dates of a Bond in years
}
