\name{fGetTimesLibor}
\alias{fGetTimesLibor}
\title{Extract the payment date of a LIBOR agreement in years}
\usage{
  fGetTimesLibor(dfInstrument)
}
\arguments{
  \item{dfInstrument}{A dataframe of instuments with at
  least columns Type and Tenor}
}
\description{
  Extract the payment date of a LIBOR agreement in years
}

