\name{fGetTimesSwap}
\alias{fGetTimesSwap}
\title{Extract the payment dates of a Swap agreement in years}
\usage{
  fGetTimesSwap(dfInstrument)
}
\arguments{
  \item{dfInstrument}{A dataframe of instuments with at
  least columns Type and Tenor}
}
\description{
  Extract the payment dates of a Swap agreement in years
}

