\name{fWilson}
\alias{fWilson}
\title{Wilson function}
\usage{
  fWilson(t, u, ufr, alpha)
}
\arguments{
  \item{t}{a time}

  \item{u}{another time}

  \item{ufr}{the ultimate forward rate}

  \item{alpha}{the spead of reversion to the ultimate
  forward rate}
}
\description{
  Acts as a kernel for regression
}

